obsolete IConfigureComponents
